% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_french_data.R
\name{browse_details_page}
\alias{browse_details_page}
\title{Browse the details page of a Kenneth's French data set}
\usage{
browse_details_page(fds)
}
\arguments{
\item{fds}{an object of class \code{french_dataset}}
}
\description{
Opens the details page of a data set on the default browser.
}
\examples{

\dontrun{
   ff_3f <- download_french_data('Fama/French 3 Factors')
   browse_details_page(ff_3f)
}

}
