% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french_data_list.R
\name{french_data_list-class}
\alias{french_data_list-class}
\title{\code{french_data_list} S3 class}
\description{
The \code{french_data_list} exists to hold the results of reading the files lists of Kenneth's French data library.

It provides a method to print the objects of this class.
}
\section{Properties of \code{french_data_list}}{


Objects of class \code{french_data_list} have:
\itemize{
\item A \code{class} attribute of \code{french_data_list}.
\item A base type of \code{"list"} with the following elements:
\itemize{
\item \code{info} - holds the information about when and were the information was retrieved.
\item \code{files_list} a tibble with with a list of files that can be downloaded, the tibble contains a \code{name}, \code{file_url} and \code{details_url} column.
}
}
}

\section{Behavior of \code{french_data_list}}{

\itemize{
\item Pretty prints the object.
}
}

