% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_french_data.R
\name{french_dataset-class}
\alias{french_dataset-class}
\title{\code{french_dataset} S3 class}
\description{
The \code{french_dataset} exists to hold the results of reading the files lists of Kenneth's French data library.

It provides a method to print the objects of this class.
}
\section{Properties of \code{french_dataset}}{


Objects of class \code{french_dataset} have:
\itemize{
\item A \code{class} attribute of \code{french_dataset}.
\item A base type of \code{"list"} with the following elements:
\itemize{
\item \code{info} - holds the information about when and were the information was retrieved.
\item \code{details_url} - url for the webpage with details on data set.
\item \code{subsets} a tibble with with the subsets contained in the downloaded file, the tibble contains a \code{name} and \code{data} column.
}
}
}

\section{Behavior of \code{french_dataset}}{

\itemize{
\item Pretty prints the object.
}
}

