% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french_data_list.R
\name{print.french_data_list}
\alias{print.french_data_list}
\title{Generic print method for objects of class \code{french_data_list}}
\usage{
\method{print}{french_data_list}(x, ...)
}
\arguments{
\item{x}{an object of class \code{french_data_list}}

\item{...}{other arguments passed to \code{print()}}
}
\description{
Prints an object of class \code{french_data_list}
}
\examples{

\donttest{
   files_list <- get_french_data_list()
   files_list
}
}
