% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french_data_list.R
\name{browse_french_site}
\alias{browse_french_site}
\title{Browse Kenneth's French data library website}
\usage{
browse_french_site()
}
\value{
Does not return a value. Opens the Kenneth's French data library website with the default browser.
}
\description{
Opens the data library website on the default browser
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\examples{
\donttest{
browse_french_site()
}

}
