% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchdata-package.R
\docType{package}
\name{frenchdata-package}
\alias{frenchdata}
\alias{frenchdata-package}
\title{frenchdata: Download Data Sets from Kenneth's French Finance Data Library Site}
\description{
Download data sets from Kenneth's French finance data library site <http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html>, reads all the data subsets from the file. Allows R users to collect the data as
    'tidyverse'-ready data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nareal.github.io/frenchdata/}
  \item \url{https://github.com/nareal/frenchdata}
  \item Report bugs at \url{https://github.com/nareal/frenchdata/issues}
}

}
\author{
\strong{Maintainer}: Nelson Areal \email{nareal@gmail.com} (\href{https://orcid.org/0000-0002-1157-0178}{ORCID})

}
\keyword{internal}
