% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_french_data.R
\name{print.french_dataset}
\alias{print.french_dataset}
\title{Generic print method for objects of class \code{french_dataset}}
\usage{
\method{print}{french_dataset}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link[=french_dataset-class]{french_dataset}}}

\item{...}{other arguments passed to \code{print()}}
}
\value{
Return a \code{tbl_df}/\code{data.frame} class object from the \code{subsets} element of \code{x}.
}
\description{
Prints an object of class \code{\link[=french_dataset-class]{french_dataset}}
}
\examples{
\donttest{
ff_3f <- download_french_data('Fama/French 3 Factors')
print(ff_3f)
ff_3f
}
}
